requirements.txt
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.0
tkinter
torch==2.0.1
torchvision==0.15.2
transformers==4.35.0
requests==2.31.0